library: replace grow_array with ASS_REALLOC_ARRAY
The latter has overflow protections for size_t indexation and element
size > 1, the former none whatsoever. Change all related vars to size_t.
Additionally, switch to exponential buffer growth.

+++ libass/ass_library.h
@@ -33,7 +33,7 @@ struct ass_library {
     char **style_overrides;
 
     ASS_Fontdata *fontdata;
-    int num_fontdata;
+    size_t num_fontdata;
     void (*msg_callback)(int, const char *, va_list, void *);
     void *msg_callback_data;
 };
